Added retrofit and dbflow modules and removed the dependencies from the core library
