+++ app/build.gradle
@@ -219,6 +219,7 @@ dependencies {
     testCompile ('org.threeten:threetenbp:1.3.2') {
         exclude group:'com.jakewharton.threetenabp', module:'threetenabp'
     }
+    compile project(path: ':retrofit')
 }
 
 //apply plugin: 'com.google.gms.google-services' // TODO: enable when adding google-services.json
