+++ app/src/androidTest/java/com/xmartlabs/template/TestComponent.java
@@ -5,7 +5,7 @@ import com.xmartlabs.base.core.module.GsonModule;
 import com.xmartlabs.base.core.module.LoggerModule;
 import com.xmartlabs.base.core.module.OkHttpModule;
 import com.xmartlabs.base.core.module.PicassoModule;
-import com.xmartlabs.base.core.module.RestServiceModule;
+import com.xmartlabs.base.retrofit.module.RestServiceModule;
 import com.xmartlabs.template.module.ControllerModule;
 import com.xmartlabs.template.ui.common.BaseInstrumentationTest;
