+++ app/src/androidTest/java/com/xmartlabs/template/module/MockRestServiceModule.java
@@ -2,7 +2,7 @@ package com.xmartlabs.template.module;
 
 import android.content.Context;
 
-import com.xmartlabs.base.core.module.RestServiceModule;
+import com.xmartlabs.base.retrofit.module.RestServiceModule;
 
 import io.appflate.restmock.RESTMockServer;
 import okhttp3.HttpUrl;
