+++ app/src/main/java/com/xmartlabs/template/ApplicationComponent.java
@@ -1,15 +1,14 @@
 package com.xmartlabs.template;
 
-import com.xmartlabs.base.core.CoreComponent;
 import com.xmartlabs.base.core.controller.Controller;
+import com.xmartlabs.base.core.controller.SharedPreferencesController;
 import com.xmartlabs.base.core.module.CoreAndroidModule;
 import com.xmartlabs.base.core.module.GsonModule;
 import com.xmartlabs.base.core.module.LoggerModule;
 import com.xmartlabs.base.core.module.OkHttpModule;
 import com.xmartlabs.base.core.module.PicassoModule;
-import com.xmartlabs.base.core.module.RestServiceModule;
+import com.xmartlabs.base.retrofit.module.RestServiceModule;
 import com.xmartlabs.template.controller.SessionController;
-import com.xmartlabs.base.core.controller.SharedPreferencesController;
 import com.xmartlabs.template.module.ControllerModule;
 import com.xmartlabs.template.ui.BaseActivity;
 import com.xmartlabs.template.ui.BaseAppCompatActivity;
@@ -26,13 +25,13 @@ import dagger.Component;
 @Component(modules = {
     CoreAndroidModule.class,
     ControllerModule.class,
-    LoggerModule.class,
     GsonModule.class,
+    LoggerModule.class,
     OkHttpModule.class,
     PicassoModule.class,
     RestServiceModule.class,
 })
-public interface ApplicationComponent extends CoreComponent {
+public interface ApplicationComponent {
   void inject(BaseProjectApplication baseProjectApplication);
 
   void inject(BaseActivity baseActivity);
