+++ core/src/main/AndroidManifest.xml
@@ -7,8 +7,6 @@
         android:allowBackup="true"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        >
-
-    </application>
+        />
 
 </manifest>
