--- core/src/main/java/com/xmartlabs/base/core/CoreComponent.java
@@ -1,25 +1,0 @@
-package com.xmartlabs.base.core;
-
-import com.xmartlabs.base.core.module.CoreAndroidModule;
-import com.xmartlabs.base.core.module.GsonModule;
-import com.xmartlabs.base.core.module.LoggerModule;
-import com.xmartlabs.base.core.module.OkHttpModule;
-import com.xmartlabs.base.core.module.PicassoModule;
-import com.xmartlabs.base.core.module.RestServiceModule;
-
-import javax.inject.Singleton;
-
-import dagger.Component;
-
-@Singleton
-@Component(modules = {
-    CoreAndroidModule.class,
-    LoggerModule.class,
-    GsonModule.class,
-    OkHttpModule.class,
-    PicassoModule.class,
-    RestServiceModule.class,
-})
-public interface CoreComponent {
-
-}
