--- core/src/main/java/com/xmartlabs/base/core/service/common/GsonRequired.java
@@ -1,13 +1,0 @@
-package com.xmartlabs.base.core.service.common;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-/** Indicates that a field must be present and have a value when deserialized using Gson. */
-@Retention(RUNTIME)
-@Target(FIELD)
-public @interface GsonRequired {
-}
