+++ dbflow/src/main/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.xmartlabs.base.dbflow"
+    >
+
+    <application
+        android:allowBackup="true"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        />
+
+</manifest>
