+++ retrofit/src/main/java/com/xmartlabs/base/retrofit/common/GsonRequired.java
@@ -1,0 +1,13 @@
+package com.xmartlabs.base.retrofit.common;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+/** Indicates that a field must be present and have a value when deserialized using Gson. */
+@Retention(RUNTIME)
+@Target(FIELD)
+public @interface GsonRequired {
+}
